Refactor and Reorg src/renderer
