+++ src/renderer/index.html
@@ -235,6 +235,6 @@
       </div>
     </div>
     
-    <script src="../renderer/renderer.js"></script>
+    <script type="module" src="./renderer.js"></script>
   </body>
 </html>
